+++ src/components/achievements/achievementCard/AchievementCard.stories.mdx
@@ -1,6 +1,6 @@
 import { Canvas, Meta, Story } from '@storybook/addon-docs';
 import AchievementCard from './AchievementCard';
-import { AchievementState, AchievementType, ActionTypes } from '../types';
+import { AchievementState, AchievementType, ActionTypes } from '../../../types/achievement';
 import KanufahrtDschungel from '../../../assets/images/achievements/KanufahrtDschungel.png';
 import SequenceExample01 from '../../../assets/images/achievements/Sequence_Example_01.png';
 import SequenceExample04 from '../../../assets/images/achievements/Sequence_Example_04.png';
