+++ src/components/achievements/achievementCard/AnimatedShine.tsx
