+++ src/components/achievements/cosmetics/AnimatedShine.stories.mdx
@@ -1,6 +1,6 @@
 import { Canvas, Meta, Story } from '@storybook/addon-docs';
 import AnimatedShine from './AnimatedShine';
-import { ShineSize } from '../types';
+import { ShineSize } from '../../../types/achievement';
 
 <Meta title="Achievements/AnimatedShine" component={AnimatedShine} />
