+++ src/components/achievements/modals/AchievementModal.stories.mdx
@@ -1,6 +1,6 @@
 import { Meta, Story } from '@storybook/addon-docs';
 import AchievementModal from './AchievementModal';
-import { AchievementState, AchievementType } from '../types';
+import { AchievementState, AchievementType } from '../../../types/achievement';
 import KanufahrtDschungel from '../../../assets/images/achievements/KanufahrtDschungel.png';
 import ExampleStreakImageRecord from '../../../assets/images/achievements/example_streak_image_record.png';
 import ExampleStreakImage from '../../../assets/images/achievements/example_streak_image.png';
