+++ src/components/achievements/polaroid/CompletePolaroid.stories.mdx
