+++ src/components/achievements/polaroid/EmptyPolaroidField.stories.mdx
