+++ src/components/achievements/progressIndicators/IndicatorBarWithSteps.stories.mdx
@@ -1,6 +1,6 @@
 import { Canvas, Meta, Story } from '@storybook/addon-docs';
 import IndicatorBarWithSteps from './IndicatorBarWithSteps';
-import { AchievementState } from '../types';
+import { AchievementState } from '../../../types/achievement';
 
 <Meta title="Achievements/ProgressIndicators/IndicatorBarWithSteps" component={IndicatorBarWithSteps} />
