+++ src/components/achievements/progressIndicators/IndicatorBarWithSteps.tsx
@@ -1,6 +1,6 @@
 import { Box, Progress, Stack, useBreakpointValue } from 'native-base';
 import IndicatorStep from './IndicatorStep';
-import { AchievementState } from '../types';
+import { AchievementState } from '../../../types/achievement';
 
 type IndicatorBarWithStepsProps = {
     maxSteps: number;
