+++ src/components/achievements/streak/StreakCard.stories.mdx
@@ -1,6 +1,6 @@
 import { Canvas, Meta, Story } from '@storybook/addon-docs';
 import StreakCard from './StreakCard';
-import { ActionTypes } from '../types';
+import { ActionTypes } from '../../../types/achievement';
 import ExampleStreakUnfinished from '../../../assets/images/achievements/example_streak_image.png';
 import ExampleStreakFinished from '../../../assets/images/achievements/example_streak_image_record.png';
