+++ src/components/achievements/streak/StreakImage.stories.mdx
@@ -1,7 +1,7 @@
 import { Canvas, Meta, Story } from '@storybook/addon-docs';
 import StreakImage from './StreakImage';
 import StreakImageContainer from './StreakImageContainer';
-import { AchievementState, StreakImageSize } from '../types';
+import { AchievementState, StreakImageSize } from '../../../types/achievement';
 import ExampleStreakUnfinished from '../../../assets/images/achievements/example_streak_image.png';
 import ExampleStreakFinished from '../../../assets/images/achievements/example_streak_image_record.png';
